HDDS-6229. [Ozone-Streaming] Data Channel abstraction on datanode (#3023)
(cherry picked from commit 3cc2a7e21be12d0a88897a4a95fc5f80f37cd100)
(cherry picked from commit de690f6c259437bdce4389c7b06b189fa4cd20c0)
(cherry picked from commit 3dde9e27b916daf73c80c5304a61477683ce4d21)
